Added a yellow enemy, they chase you.
Need to make sure that animation and coroutine is done when hit to make them stop moving and play animation.
