+++ Assets/Enemies/Animations/Running.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 716f02416c4efa24eae14bc470e8f15a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
