+++ Assets/Enemies/Models/Beta_Joints_MAT.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c833de5c962425445ad461180b05f6c0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
