+++ Assets/Enemies/Models/Joints.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 765464c3edfb85444a732fb43e0455f2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
