+++ Assets/Enemies/Models/Limbs.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 888d3b1f771f08d4387f3472af17d085
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
