+++ Assets/Enemies/Models/asdf1_Beta_HighLimbsGeoSG2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 86da718c5afbff34eb0d07271ae502a1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
